+++ src/emu/romentry.h
@@ -132,18 +132,8 @@ struct tiny_rom_entry
 class rom_entry
 {
 public:
-	rom_entry(const tiny_rom_entry &ent)
-		: m_name(ent.name != nullptr ? ent.name : "")
-		, m_hashdata(ent.hashdata != nullptr ? ent.hashdata : "")
-		, m_offset(ent.offset)
-		, m_length(ent.length)
-		, m_flags(ent.flags) {}
-	rom_entry(std::string &&name, std::string &&hashdata, UINT32 offset, UINT32 length, UINT32 flags)
-		: m_name(std::move(name))
-		, m_hashdata(std::move(hashdata))
-		, m_offset(offset)
-		, m_length(length)
-		, m_flags(flags) {}
+	rom_entry(const tiny_rom_entry &ent);
+	rom_entry(std::string &&name, std::string &&hashdata, UINT32 offset, UINT32 length, UINT32 flags);
 	rom_entry(rom_entry const &) = default;
 	rom_entry(rom_entry &&) = default;
 	rom_entry &operator=(rom_entry const &) = default;
